requirement.txt
python
django
django_rest_framework
coresheaders